Split modules into three pages
