+++ frontend/package.json
@@ -10,9 +10,15 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "dayjs": "^1.10.7",
+    "http-proxy-middleware": "^2.0.2",
+    "js-cookie": "^3.0.1",
+    "jwt-check-expiration": "^1.0.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "web-vitals": "^2.1.4"
+    "react-router-dom": "^6.2.1",
+    "web-vitals": "^2.1.4",
+    "styled-components": "^5.3.3"
   },
   "devDependencies": {
     "react-scripts": "5.0.1"
