+++ frontend/src/styles/Dashboard.css
@@ -1,0 +1,10 @@
+.dashboard {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    height: 100%;
+    width: 100%;
+    background-color: #f5f5f5;
+
+}
